+++ dds/DCPS/ValueCommon.h
@@ -47,27 +47,8 @@ public:
     pairs_ = pairs;
   }
 
-  bool get_value(ACE_CDR::Long& value, const char* name) const
-  {
-    for (OPENDDS_VECTOR(Pair)::const_iterator it = pairs_.begin(); it != pairs_.end(); ++it) {
-      if (std::strcmp(it->name, name) == 0) {
-        value = it->value;
-        return true;
-      }
-    }
-    return false;
-  }
-
-  bool get_name(const char*& name, ACE_CDR::Long value) const
-  {
-    for (OPENDDS_VECTOR(Pair)::const_iterator it = pairs_.begin(); it != pairs_.end(); ++it) {
-      if (it->value == value) {
-        name = it->name;
-        return true;
-      }
-    }
-    return false;
-  }
+  bool get_value(ACE_CDR::Long& value, const char* name) const;
+  bool get_name(const char*& name, ACE_CDR::Long value) const;
 
   XTypes::TypeKind get_equivalent_int() const
   {
@@ -83,7 +64,11 @@ class BitmaskHelper {
 public:
   virtual ~BitmaskHelper() {}
   virtual bool get_value(ACE_CDR::ULongLong& value, const OPENDDS_VECTOR(const char*)& names) const = 0;
+
+  // Return an estimated length of a string constructed by the returned flag names
+  // with a delimiter character between two consecutive flags.
   virtual size_t get_names(OPENDDS_VECTOR(const char*)& names, ACE_CDR::ULongLong value) const = 0;
+
   virtual XTypes::TypeKind get_equivalent_uint() const = 0;
 };
 
@@ -99,59 +84,25 @@ public:
   typedef PosToNameMap::const_iterator ptn_iterator;
   typedef NameToPosMap::const_iterator ntp_iterator;
 
+  MapBitmaskHelper(XTypes::TypeKind as_uint)
+    : as_uint_(as_uint)
+  {}
+
   MapBitmaskHelper(ACE_CDR::UShort bit_bound, XTypes::TypeKind as_uint)
     : bit_bound_(bit_bound)
     , as_uint_(as_uint)
   {}
 
-  MapBitmaskHelper(const Pair* pairs, ACE_CDR::UShort bit_bound, XTypes::TypeKind as_uint)
-    : bit_bound_(bit_bound)
-    , as_uint_(as_uint)
+  MapBitmaskHelper(const Pair* pairs, ACE_CDR::UShort bit_bound, XTypes::TypeKind as_uint);
+  void pairs(const OPENDDS_VECTOR(Pair)& pairs);
+
+  void bit_bound(ACE_CDR::UShort bound)
   {
-    for (const Pair* ptr = pairs; ptr->name; ++ptr) {
-      pos_to_name_.insert(std::make_pair(ptr->position, ptr->name));
-      name_to_pos_.insert(std::make_pair(ptr->name, ptr->position));
-    }
+    bit_bound_ = bound;
   }
 
-  void pairs(const OPENDDS_VECTOR(Pair)& pairs)
-  {
-    for (OPENDDS_VECTOR(Pair)::const_iterator it = pairs.begin(); it != pairs.end(); ++it) {
-      pos_to_name_.insert(std::make_pair(it->position, it->name));
-      name_to_pos_.insert(std::make_pair(it->name, it->position));
-    }
-  }
-
-  bool get_value(ACE_CDR::ULongLong& value, const OPENDDS_VECTOR(const char*)& names) const
-  {
-    ACE_CDR::ULongLong rtn = 0;
-    for (size_t i = 0; i < names.size(); ++i) {
-      const ntp_iterator it = name_to_pos_.find(names.at(i));
-      if (it == name_to_pos_.end()) {
-        return false;
-      }
-      rtn |= 1 << it->second;
-    }
-    value = rtn;
-    return true;
-  }
-
-  // Return an estimated length of a string constructed by the returned flag names
-  // with a delimiter character between two consecutive flags.
-  size_t get_names(OPENDDS_VECTOR(const char*)& names, ACE_CDR::ULongLong value) const
-  {
-    size_t rtn_size = 0;
-    OPENDDS_VECTOR(const char*) rtn;
-    for (ACE_CDR::UShort i = 0; i < bit_bound_; ++i) {
-      const ptn_iterator it = pos_to_name_.find(i);
-      if ((it != pos_to_name_.end()) && (value & 1 << i)) {
-        rtn.push_back(it->second);
-        rtn_size += std::strlen(it->second) + 1; // +1 for a delimiter like a pipe ('|') character
-      }
-    }
-    names = rtn;
-    return rtn_size;
-  }
+  bool get_value(ACE_CDR::ULongLong& value, const OPENDDS_VECTOR(const char*)& names) const;
+  size_t get_names(OPENDDS_VECTOR(const char*)& names, ACE_CDR::ULongLong value) const;
 
   XTypes::TypeKind get_equivalent_uint() const
   {
